Created somehwat better structure by creating a new class to handle processing requests. The method that switches based off of "directive" (or the enum-like header) will call methods from this class statically, at least for now.  Planning on learning more about MVVM and will hopefully be able to refactor completely to MVVM standards soon.  Project is set to create logic for all the standard use cases.
